Refactor temporal mapping
Many of the estimations are made by comparing values from two dates
with various offsets.
With data managed in dictionary with Date keys, some helpers can be
made to consolidate this common logic in a consistent way, hopefully
reducing the code, improving readability, and making extensions to the
model easier.